+++ redacted-compiler-plugin/src/main/kotlin/dev/zacsweers/redacted/compiler/RedactedIrGenerationExtension.kt
@@ -19,21 +19,21 @@ import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
-import org.jetbrains.kotlin.name.FqName
+import org.jetbrains.kotlin.name.ClassId
 
 internal class RedactedIrGenerationExtension(
   private val messageCollector: MessageCollector,
   private val replacementString: String,
-  private val redactedAnnotationName: FqName,
-  private val unRedactedAnnotationName: FqName,
+  private val redactedAnnotations: Set<ClassId>,
+  private val unRedactedAnnotations: Set<ClassId>,
 ) : IrGenerationExtension {
 
   override fun generate(moduleFragment: IrModuleFragment, pluginContext: IrPluginContext) {
     val redactedTransformer =
       RedactedIrVisitor(
         pluginContext,
-        redactedAnnotationName,
-        unRedactedAnnotationName,
+        redactedAnnotations,
+        unRedactedAnnotations,
         replacementString,
         messageCollector,
       )