+++ redacted-compiler-plugin/src/main/kotlin/dev/zacsweers/redacted/compiler/fir/FirRedactedExtensionRegistrar.kt
@@ -15,6 +15,7 @@
  */
 package dev.zacsweers.redacted.compiler.fir
 
+import dev.zacsweers.redacted.compiler.firstNotNullResult
 import dev.zacsweers.redacted.compiler.unsafeLazy
 import org.jetbrains.kotlin.descriptors.isEnumEntry
 import org.jetbrains.kotlin.descriptors.isObject
@@ -31,7 +32,7 @@ import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirFunction
 import org.jetbrains.kotlin.fir.declarations.FirProperty
 import org.jetbrains.kotlin.fir.declarations.getAnnotationByClassId
-import org.jetbrains.kotlin.fir.declarations.hasAnnotation
+import org.jetbrains.kotlin.fir.declarations.toAnnotationClassIdSafe
 import org.jetbrains.kotlin.fir.declarations.utils.isEnumClass
 import org.jetbrains.kotlin.fir.declarations.utils.isExpect
 import org.jetbrains.kotlin.fir.declarations.utils.isExtension
@@ -43,6 +44,8 @@ import org.jetbrains.kotlin.fir.declarations.utils.superConeTypes
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
 import org.jetbrains.kotlin.fir.resolve.fullyExpandedType
 import org.jetbrains.kotlin.fir.resolve.toSymbol
+import org.jetbrains.kotlin.fir.symbols.SymbolInternals
+import org.jetbrains.kotlin.fir.symbols.resolvedAnnotationsWithClassIds
 import org.jetbrains.kotlin.fir.types.ConeErrorType
 import org.jetbrains.kotlin.fir.types.classId
 import org.jetbrains.kotlin.fir.types.coneType
@@ -51,11 +54,11 @@ import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.util.OperatorNameConventions
 
 internal class FirRedactedExtensionRegistrar(
-  private val redactedAnnotation: ClassId,
-  private val unRedactedAnnotation: ClassId,
+  private val redactedAnnotations: Set<ClassId>,
+  private val unRedactedAnnotations: Set<ClassId>,
 ) : FirExtensionRegistrar() {
   override fun ExtensionRegistrarContext.configurePlugin() {
-    +RedactedFirBuiltIns.getFactory(redactedAnnotation, unRedactedAnnotation)
+    +RedactedFirBuiltIns.getFactory(redactedAnnotations, unRedactedAnnotations)
     +::FirRedactedCheckers
   }
 }
@@ -70,37 +73,37 @@ internal class FirRedactedCheckers(session: FirSession) : FirAdditionalCheckersE
 
 internal object FirRedactedDeclarationChecker : FirClassChecker(MppCheckerKind.Common) {
   override fun check(declaration: FirClass, context: CheckerContext, reporter: DiagnosticReporter) {
-    val classRedactedAnnotation =
-      declaration.getAnnotationByClassId(context.session.redactedAnnotation, context.session)
-    val classIsRedacted = classRedactedAnnotation != null
-    val classUnRedactedAnnotation =
-      declaration.getAnnotationByClassId(context.session.unRedactedAnnotation, context.session)
-    val classIsUnRedacted = classUnRedactedAnnotation != null
+    val classRedactedAnnotations =
+      context.session.redactedAnnotations.mapNotNull {
+        declaration.getAnnotationByClassId(it, context.session)
+      }
+    val classIsRedacted = classRedactedAnnotations.isNotEmpty()
+    val classUnRedactedAnnotations =
+      context.session.unRedactedAnnotations.mapNotNull {
+        declaration.getAnnotationByClassId(it, context.session)
+      }
+    val classIsUnRedacted = classUnRedactedAnnotations.isNotEmpty()
     val supertypeIsRedacted by unsafeLazy {
       declaration.superConeTypes.any {
         if (it is ConeErrorType) return@any false
-        it.classId
-          ?.toSymbol(context.session)
-          ?.hasAnnotation(context.session.redactedAnnotation, context.session) == true
+        it.classId?.toSymbol(context.session)?.resolvedAnnotationClassIds?.any {
+          it in context.session.redactedAnnotations
+        } == true
       }
     }
-    var anyRedacted = false
-    var anyUnredacted = false
 
+    val redactedProperties = mutableMapOf<FirProperty, ClassId>()
+    val unredactedProperties = mutableMapOf<FirProperty, ClassId>()
     for (prop in declaration.declarations.filterIsInstance<FirProperty>()) {
-      val isRedacted = prop.isRedacted(context.session)
-      val isUnredacted = prop.isUnredacted(context.session)
-      if (isRedacted) {
-        anyRedacted = true
-      }
-      if (isUnredacted) {
-        anyUnredacted = true
-      }
+      prop.redactedAnnotation(context.session)?.let { redactedProperties[prop] = it }
+      prop.unredactedAnnotation(context.session)?.let { unredactedProperties[prop] = it }
     }
+    val anyRedacted = redactedProperties.isNotEmpty()
+    val anyUnredacted = unredactedProperties.isNotEmpty()
 
-    val redactedName = { context.session.redactedAnnotation.shortClassName.asString() }
+    val redactedName = { redactedProperties.values.first().shortClassName.asString() }
 
-    val unRedactedName = { context.session.redactedAnnotation.shortClassName.asString() }
+    val unRedactedName = { unredactedProperties.values.first().shortClassName.asString() }
 
     if (classIsRedacted || supertypeIsRedacted || classIsUnRedacted || anyRedacted) {
       val customToStringFunction =
@@ -149,16 +152,17 @@ internal object FirRedactedDeclarationChecker : FirClassChecker(MppCheckerKind.C
       }
       if (declaration.classKind.isObject) {
         if (!supertypeIsRedacted) {
+          val classAnnotation = classRedactedAnnotations.first()
           reporter.reportOn(
-            classRedactedAnnotation!!.source,
+            classAnnotation.source,
             RedactedDiagnostics.REDACTED_ERROR,
-            "@${redactedName()} is useless on object classes.",
+            "@${classAnnotation.toAnnotationClassIdSafe(context.session)?.shortClassName?.asString()} is useless on object classes.",
             context,
           )
           return
         } else if (classIsUnRedacted) {
           reporter.reportOn(
-            classUnRedactedAnnotation.source,
+            classUnRedactedAnnotations.firstOrNull()?.source,
             RedactedDiagnostics.REDACTED_ERROR,
             "@${unRedactedName()} is useless on object classes.",
             context,
@@ -200,6 +204,14 @@ internal object FirRedactedDeclarationChecker : FirClassChecker(MppCheckerKind.C
           "@${redactedName()} should only be applied to the class or its properties, not both.",
           context,
         )
+        for ((redactedProp, annotationId) in redactedProperties) {
+          reporter.reportOn(
+            redactedProp.source,
+            RedactedDiagnostics.REDACTED_ERROR,
+            "@${annotationId.shortClassName.asString()} should only be applied to the class or its properties, not both.",
+            context,
+          )
+        }
         return
       }
       // Rest filled in by the IR plugin
@@ -213,11 +225,17 @@ internal object FirRedactedDeclarationChecker : FirClassChecker(MppCheckerKind.C
       valueParameters.isEmpty() &&
       returnTypeRef.coneType.fullyExpandedType(session).isString
 
-  private fun FirProperty.isRedacted(session: FirSession): Boolean =
-    hasAnnotation(session.redactedAnnotation, session)
+  @OptIn(SymbolInternals::class)
+  private fun FirProperty.redactedAnnotation(session: FirSession): ClassId? =
+    resolvedAnnotationsWithClassIds(symbol).firstNotNullResult {
+      it.toAnnotationClassIdSafe(session).takeIf { it in session.redactedAnnotations }
+    }
 
-  private fun FirProperty.isUnredacted(session: FirSession): Boolean =
-    hasAnnotation(session.unRedactedAnnotation, session)
+  @OptIn(SymbolInternals::class)
+  private fun FirProperty.unredactedAnnotation(session: FirSession): ClassId? =
+    resolvedAnnotationsWithClassIds(symbol).firstNotNullResult {
+      it.toAnnotationClassIdSafe(session).takeIf { it in session.unRedactedAnnotations }
+    }
 
   private val FirClass.isInstantiableEnum: Boolean
     get() = isEnumClass && !isExpect && !isExternal