+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Changelog
 
 - Fix FIR diagnostics rendering in the IDE. Note this only works in the K2 Kotlin IDE plugin + setting the IntelliJ `kotlin.k2.only.bundled.compiler.plugins.enabled` registry key to `false`.
 - When custom annotations are defined, report those names in FIR error messages.
+- Support multiple custom annotations.
+  - For Gradle configuration, the singular `*Annotation` properties are deprecated in favor of plural`*Annotations` `SetProperty` types.
+  - For CLI consumers, the `redactedAnnotation` and `unredactedAnnotation` properties are now `redactedAnnotations` and `unredactedAnnotations`.
 - Build against Gradle `8.12`.
 - Only report errors in FIR now. Removes the `validateIr` plugin option.
 - No longer support K1.